Fix usage of absolute where relative should be
## Problem
The example in "Use absolute path for internal links" section represents absolute path to be a bad practice but the heading and the text of the section say otherwise. Also, for internal links relative path is better as the protocol and domain name can be omitted as they already match the site's.

## Fix
I've made changes by replacing absolute with relative in the section.